refactor(python): extract exception hierarchy to mrrc/exceptions.py
bd-qzew added ~380 lines of exception class hierarchy to mrrc/__init__.py,
pushing it past the threshold where the structure starts hurting (mixing
re-exports, module-level helpers, exception classes, format constants,
and several large wrapper classes). The exception hierarchy is the
easiest unit to extract cleanly since nothing else in the file depends
on it at import time.

Moves _MrrcExceptionBase, _POSITIONAL_FIELDS, MrrcException, all 12
mrrc-named exception classes (RecordLengthInvalid, RecordLeaderInvalid,
BaseAddressInvalid, BaseAddressNotFound, RecordDirectoryInvalid,
EndOfRecordNotFound, FieldNotFound, FatalReaderError, InvalidIndicator,
BadSubfieldCode, InvalidField, TruncatedRecord, EncodingError, XmlError,
JsonError, WriterError), and BadSubfieldCodeWarning to a new
mrrc/exceptions.py module.

mrrc/__init__.py re-exports all of them so the public API
(mrrc.InvalidIndicator, mrrc.RecordLeaderInvalid, etc.) is byte-for-byte
identical. Aligns with pymarc's pymarc/exceptions.py convention.

mrrc/__init__.py reduced from 2339 to 1971 lines (-368 net after the
imports added back). All 647 Python tests pass unmodified.

Filed bd-mcei for the larger refactor of the remaining wrapper classes
(Field, Record, Leader, MARCReader, MARCWriter) which together are
~1500 more lines and would naturally live in their own modules
following the same pymarc-aligned pattern.
