+++ .github/workflows/python-release.yml
@@ -122,11 +122,11 @@ jobs:
       - name: Install wheel and test dependencies
         shell: bash
         run: |
-          # Install the wheel together with the [test] extra from
-          # pyproject.toml so test deps stay in sync with the package.
+          # Install the wheel together with the [dev] extra (which extends
+          # [test] via self-reference) so test runtime deps + type checkers
+          # stay in sync with pyproject.toml as a single source of truth.
           WHEEL=$(ls dist/mrrc-*.whl | head -1)
-          uv pip install --system "${WHEEL}[test]"
-          uv pip install --system mypy pyright
+          uv pip install --system "${WHEEL}[dev]"
 
       - name: Run pytest (core tests, excludes benchmarks)
         run: |
