+++ pyproject.toml
@@ -27,19 +27,19 @@ classifiers = [
 ]
 
 [project.optional-dependencies]
-# Runtime test dependencies — kept as a separate extra so the wheel-test CI
-# workflows can install them via `mrrc[test]` and stay in sync with this list
-# automatically (no parallel hardcoded list to keep updated).
+# Runtime test dependencies — used by the wheel-test CI workflows via
+# `mrrc[test]` so the dep list has a single source of truth and any future
+# test dep added here is automatically picked up by CI.
 test = [
     "pytest>=7.0",
     "pytest-benchmark>=4.0",
     "syrupy>=4.0",
 ]
 # Developer tooling (linters, type checkers) on top of the test deps.
+# Self-references `mrrc[test]` so the test deps stay listed in exactly one
+# place above; pip and uv both resolve self-referential extras correctly.
 dev = [
-    "pytest>=7.0",
-    "pytest-benchmark>=4.0",
-    "syrupy>=4.0",
+    "mrrc[test]",
     "mypy>=1.0",
     "pyright>=1.1",
     "ruff>=0.1",
